scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.2",
    "azure-cosmos>=4.14.2",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
